CORE-731 Move app launch stories into separate files.
