+++ stories/apps/launch/data/JupyterLabNoParamsApp.js
